modify style_repaint
