+++ modelscope_agent/tools/style_repaint.py
@@ -1,10 +1,9 @@
 import os
 import time
-from configparser import ConfigParser
 
 import json
-import oss2
 import requests
+from modelscope_agent.tools.localfile2url_utils.localfile2url import path2url
 from modelscope_agent.tools.tool import Tool, ToolSchema
 from pydantic import ValidationError
 from requests.exceptions import RequestException, Timeout
@@ -12,48 +11,6 @@ from requests.exceptions import RequestException, Timeout
 MAX_RETRY_TIMES = 3
 
 WORK_DIR = os.getenv('CODE_INTERPRETER_WORK_DIR', '/tmp/ci_workspace')
-
-
-def upload_to_oss(bucket, local_file_path, oss_file_path):
-    # 上传文件到阿里云OSS
-    bucket.put_object_from_file(oss_file_path, local_file_path)
-
-    # 设置文件的公共读权限
-    bucket.put_object_acl(oss_file_path, oss2.OBJECT_ACL_PUBLIC_READ)
-
-    # 获取文件的公共链接
-    file_url = f"https://{bucket.bucket_name}.{bucket.endpoint.replace('http://', '')}/{oss_file_path}"
-    return file_url
-
-
-def get_oss_config():
-    # 尝试从环境变量中读取配置
-    access_key_id = os.getenv('OSS_ACCESS_KEY_ID')
-    access_key_secret = os.getenv('OSS_ACCESS_KEY_SECRET')
-    endpoint = os.getenv('OSS_ENDPOINT')
-    bucket_name = os.getenv('OSS_BUCKET_NAME')
-
-    # 如果环境变量没有设置，尝试从.ossutilconfig文件中读取
-    if not access_key_id or not access_key_secret or not endpoint or not bucket_name:
-        config = ConfigParser()
-        config.read(os.path.expanduser('~/.ossutilconfig'))
-        if 'Credentials' in config:
-            access_key_id = config.get('Credentials', 'accessKeyId')
-            access_key_secret = config.get('Credentials', 'accessKeySecret')
-            endpoint = config.get('Credentials', 'endpoint')
-            bucket_name = config.get('Credentials', 'bucketName')
-
-    return access_key_id, access_key_secret, endpoint, bucket_name
-
-
-def path2url(local_file_path, oss_file_path):
-    local_file_path = os.path.join(WORK_DIR, local_file_path)
-    ak_id, ak_secret, endpoint, bucket_name = get_oss_config()
-    auth = oss2.Auth(ak_id, ak_secret)
-    bucket = oss2.Bucket(auth, endpoint, bucket_name)
-    file_url = upload_to_oss(bucket, local_file_path,
-                             f'agents/user/{oss_file_path}')
-    return file_url
 
 
 class StyleRepaint(Tool):
@@ -153,10 +110,12 @@ class StyleRepaint(Tool):
                 restored_dict[key] = value
             kwargs = restored_dict
             image_path = kwargs['input'].pop('image_path', None)
-            if image_path:
+            if image_path and image_path.endswith(('.jpeg', '.png', '.jpg')):
                 # 生成 image_url，然后设置到 kwargs['input'] 中
                 image_url = path2url(image_path, f'{self.name}/{image_path}')
                 kwargs['input']['image_url'] = image_url
+            else:
+                raise ValueError('请先上传一张正确格式的图片')
             kwargs['model'] = 'wanx-style-repaint-v1'
             # kwargs['input']['style_index'] = int(kwargs['input']['style_index'])
         print('传给style_repaint tool的参数：', kwargs)
